Backend/UserApp/serializers.py
import imp
from rest_framework import serializers
from UserApp.models import Usuario

class UsuarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Usuario
        fields = '__all__'